+++ lib/vapor_sodium.h
@@ -1,6 +1,6 @@
 ﻿#pragma once
 
-using data_type = float;
+using data_type = double;
 
 /**
  * @brief Provides thermophysical and transport properties for Sodium Vapor.
